Incomplete Analysis
Analysis of sever-side uploads is done.
I added two more csv files, one for keeping track of download rates and one for response times. Just the values are needed for this, keeping track of the files involved in interactions is unnecessary. The implementation of updating these files is incomplete though I have started on that.